pyproject.toml
[tool.poetry]
name = "telegram-chatgpt"
version = "0.1.0"
description = ""
authors = ["edoardob90 <edoardob90@gmail.com>"]
readme = "README.md"
packages = [{include = "telegram_chatgpt"}]

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.1"
python-dotenv = "^1.0.0"
openai = "^0.27.0"
tiktoken = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
